Moving createClient to another file and adding validation for agent to ensure details are there
