+++ server/services/agentService.js
@@ -1,0 +1,5 @@
+module.exports = {
+  agentRequiresSetup: agent => {
+    return !agent.firstName || !agent.lastName || !agent.businessName || !agent.abn || !agent.phone;
+  }
+};
